Update patch set 2
Patch Set 2: Code-Review-1

This isn't going to work, as rdo-jobs is an untrusted project and we don't let check and gate pipelines use secrets.  We'll create a new legacy job in config which will be able to run in check and gate, also allows us to lock down who can propose changes and access the secret.

Patch-set: 2
Label: Code-Review=-1