Get the full list of old subpackages in rpmlint job
Currently, in rpmlint job, the list of old packages is taken
asuming that is the same subpackages as in new packages. This
makes rpmlint job to fail in cases where we are adding or removing
packages [1].

This job is looking for the list of old packages by querying the
packages with the same srpm.

[1] https://review.rdoproject.org/r/#/c/14588/

Change-Id: I60f20c6f44616ee84fe91a148822507b2ecf1f90