Update patch set 9
Patch Set 9: Code-Review+1

lgtm, but have not tested - following the work.  

If we had a list of the jobs using the nodesets defined here it might make for a nicer commit message, or perhaps this is better suited to a different commit (to a readme).  In any event as this work is already in flight it's just an idea.

Patch-set: 9
Reviewer: Gerrit User 202 <202@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1