Add first job reparented to tripleo-ci-base - Browbeat
This review adds a first job based off the tripleo-ci-base
base job and uses the templated toci scripts. The playbooks
override is added to this job definition - for when
toci_gate_test will start using it.

The companion legacy job will run concurrently for now.

As the playbooks for ovb have been moved, the following
patch needs to land prior to this:

https://review.openstack.org/590090 Add browbeat-minimal.yml to ovb playbook lis

Change-Id: Ia51799e7b40195f081830afed8dccfb40e1a19d9
Depends-On: https://softwarefactory-project.io/13474/
Depends-On: https://review.openstack.org/590090
Depends-On: https://review.openstack.org/#/c/587228/