Add container nodeset
This is to make the base job usable on nodes that doesn't have sudo.
This change adds new container nodeset to be able to use faster and
thiner node for simple jobs like tox. To do that, we need to update
the configure-mirrors tasks to only run when the user can sudo.

This change also update the rdo-jobs tox-docs job to run with the new
nodeset.

Change-Id: I93706c76f98a6117a3bde76a1bd23c14de8f5909