Disable vxlan_networking for upgrades jobs.
Currently, this check is failing in all mixed_upgrades jobs,
as the respositories are not being set up in the subnode-1
due to different releases being deployed in the undercloud
and overcloud. This causes having an old ovs version installed
in the subnode that makes the vxlan_networking script fail.

Change-Id: Icd2fa17aa681c309f079fd3beca9bed57bc6453f
Closes-Bug: https://bugs.launchpad.net/tripleo/+bug/1791115