Add DLRN-rpmbuild-pinned and DLRN-rpmbuild-fedora-pinned
Sometimes, specs in rpm-master branches don't work fine with master
branches of upstream projects. In those cases we can't test changes
in rpm-master branches properly as we need to test them using the
pinned versions in rdoinf tags.

This patch is adding two new jobs DLRN-rpmbuild-pinned and
DLRN-rpmbuild-fedora-pinned which are similar to DLRN-rpmbuild
and DLRN-rpmbuild-fedora but using rdoinfo stein-uc and stein-py3-uc
tags to find out what branches, tags or commits to use for the job.

Change-Id: Id9164634b3646e8a075b52f003992ff35f192c24