Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Some suggestions (inline) to reduce redundancy.

Ideally, we would not need the *-pinned/run.yaml playbooks, since they just clone the non-pinned playbooks and add a tag in one line. We could just add a variable to set the tag in all jobs, then make it empty for non-pinned playbooks and set to a value for the pinned ones.

Patch-set: 2
Label: Code-Review=-1