Update patch set 14
Patch Set 14:

Félix Enrique Llorente Pastora wrote:

> The base jobs are not already defined in the Depends-On ?

Yes, but the depends-On is pro-forma, since it's on a trusted project, it cannot really be tested before it's merged, so to test the translated job right noe, we need to duplicate definition here. After the Depends-On is merged, we can remove the definitions from here, and test again that they are working.

Patch-set: 14
Label: Code-Review=0