Update patch set 1
Patch Set 1: Code-Review-1

we have a bug in the script that affects one corner case. If in the same review we update common-testing and rocky-testing or queens-testing. The queens or rocky job will enable the temporary common-testing repo which is incorrect as we don't use common since queens.

Patch-set: 1
Label: Code-Review=-1