Add new job rdoinfo-tripleo-master-testing-centos-7-ovb-3ctlr_1comp-featureset030
To properly gate OVS updates, running an OVB multinode job using
FS030 (based on OVN driver) is recommended. This job is adding an
ovb job running that featureset. This job will be included in
experimental pipeline until we have a better way of assigning specific
jobs to packages (based on rdoinfo reorganization).

Change-Id: Iece85581e53f8766c6c29cd319a7658e76a50b74