Skip weirdo executions if buildset_artifacts_url is undefined
In weirdo jobs nested in dlrn build jobs, if the parent DLRN-rpmbuild
job has not created any new package, buildset_artifacts_url variable is
not defined and weirdo jobs fail.

This patch is creating a block to to check if buildset_artifacts_url is
defined. If it's undefined it will just skip weirdo execution.

Change-Id: I9ef55da710dcf1aa896b521be3e573b1ed05403b