Add arch to job naming to allow for future job additions
The inclusion of Fedora-28 in the promotion pipelines
required the expansion of the use of jobs names over job_id
in the promotion criteria.

Along with this change, this review renames the jobs in rdocloud
to include the arch on which the job is run in prepartion
for future expansion of jobs arch selection.

Change-Id: Ia263a92da31ba5c636680a0afb6bca0c043249f1