Adds tripleo-build-containers-fedora-28 non voting
We want to run the tripleo-build-containers-fedora-28 as non voting jobs
on openstck-infra/tripleo-ci changes for that we need to create a new
project-template just including this job.

The name of the project template is the same as the upstream one so it's
easier to map, or incase we add other distros jobs (future centos8) we
use the same.

Story: https://tree.taiga.io/project/tripleo-ci-board/task/725?kanban-status=1447276

Change-Id: Ia86baa128db4dab2b87f3aeaf2d1c5cda07f4cf2