Enable containers update when building containers
We have discovered that oooq jobs with rdoinfo_build_containers are not
using the containers just built in the job but the promoted ones. This
is making oooq deployments meaningless in terms of test coverage.

This patch enables update_containers step in oooq when enabled
containers build so that at least we test the updated packages in
rdoinfo.

Note that, once we can use the containers created in the job for the
oooq deployment we can disable update_containers again.

Change-Id: I873db1d501d38aa8aa33b4f9fe1e43a1e689dc63