Port standalone scenario01 to os_tempest
In order to use os_tempest ansible tempest role
we need to port the featureset overrides in periodic
standalone scenario01 so that it will consume the os_tempest
role.

Currently heat is not supported in tempestconf, so we
are removing it.

Add Member role to tempestconf extra config and tempest_worker
is not needed as tests runs serially in os_tempest.

https://tree.taiga.io/project/tripleo-ci-board/task/906

Depends-On: https://review.openstack.org/#/c/650243/
Depends-On: https://review.openstack.org/649577
Depends-On: https://review.openstack.org/649553

Change-Id: Ic79316c4a6ff812504a2ef1f8e929ee5d2d34b7e
Signed-off-by: Chandan Kumar <chkumar246@gmail.com>