[DNM] Switch to buildah based jobs
We need to switch away from rdo-kolla-build role
as there is no support for buildah in kolla, so
let's start using tripleo-buildcontainers playbooks
from tripleo-ci.

To reuse tasks from upstream tripleo-buildcontainers
need to be switched to roles, then we can reuse it here,
for now starting with copying and adjusting playbooks
for our use.

Change-Id: I2d914ab41a3d8b7ddf6e4e64aad1daf14e337fd6