Add new fs020/021 jobs to use 2compute nodes
https://review.opendev.org/#/c/660710/ adds 1ctrl_2comp nodes,
in order to improve the testing performance of fs020, we are
adding more nodes.

We cannot update the existing job name as it will break at
multiple places in promotion pipeline. So we are adding
the new set of jobs then once these jobs gets added, we can
remove the 1comp fs020/21 jobs.

We donot need to use branches tag in zuul job definition as those
fs020/fs021 check jobs runs against branchless project where master
branch exists and zuul is going to take a care of that.

https://tree.taiga.io/project/tripleo-ci-board/task/1080

Change-Id: I126143a4b52063420c44d38aca2090ca3dc2f59a
Signed-off-by: Chandan Kumar (raukadah) <chkumar246@gmail.com>