Update patch set 2
Patch Set 2:

> (1 comment)

Although that might be the issue, I prefer the second approach with inheritance, I know it's annoying have to add the ansible_python_interpreter in all these jobs, but once we move to ansible 2.8 it will be just a matter of remove the line.
Another approach would be create base-featureset00[1-4], and have centos-7 and rhel-8 jobs inherit from these base-featureset00[1-4] jobs.

Patch-set: 2