Change default ansible to 2.8 in base job
Please use `Depends-On:` with this change in order to verify that your
jobs can deal with this newer version of Ansible.

If they do not please update them and manually set ansible-version
at their level, so they will not break when we merge this change.

This also removes base-rolling job based on feedback we received from
pbelanger which recommended this new approach which avoids multiple
base jobs. He already used the same approach with Ansible Zuul and it
went well for him.

Change-Id: Ia06b9f6b624180ba8a3e9827f407a1972f6b70c2
Task: https://tree.taiga.io/project/tripleo-ci-board/task/1256