Define periodic jobs for the train branch
Since stable/train branch is branched from master, for job
parenting, Master job as a parent is used for Train by
overriding the branch and release name.

Since FS016,17,18,19 does not run in promotion pipeline as
they are replaced with standalone scenario-1,2,3,4 so jobs
are not defined for train

For train RHEL-8 jobs, ansible_python_interpreter is not needed
as ansible-2.8 auto detects it.

Story: https://tree.taiga.io/project/tripleo-ci-board/us/1288
Change-Id: Ic2070353d741c7366b09cbd80bcec9b193af618c