Add reqcheck job in distgit project pipeline
We run "rdopkg reqcheck" in a job as non-voting (for the moment)
to check if dependencies are matching between spec and requirements
file.

We run the job in a fedora container, but we can go a step further
by using a dedicated container for this job with the last rdopkg
version. And rebuild this container each time a new rdopkg version
is released in pip. It would be time-saving.

Change-Id: I2b12f124660e94646ed2a4bc74f30fb9a2ed0a95