Improve ceph-ansible jobs
This patch remove an old featureset016 that is no longer required on
ceph-ansible, and also remove tempest settings which are not used in
ceph-ansible, since the purpose of these jobs are to test the
ceph-ansible role, not tempest.
The patch also makes a better use of yaml anchors for future periodic
jobs.

Change-Id: Id1168d302cf6de1800f70dd06ac6b95b75fb8dbc