Move Create repo task to run to avoid post faiure
Currently in podman build rpm job, if build rpm task fails
then the create repo task is called in post playbook where rpm is
not found then it leads to post failure as post playbook failed.

In order to avoid, we are moving the build rpm task in run playbook
so that the job will return failure if the task fails.

Change-Id: Iecdb11a236db6a87309bc4b8081d102b26cc0aaa
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>