Update patch set 9
Patch Set 9: Code-Review-1

Overall LGTM but is missing one important bit: it fails to trigger the jobs defined in ceph-ansible.yaml file when the file is touched, risking to break them on changes.

How about triggering these when the file is touched? (so we do *not* need to create extra test changes)

Patch-set: 9
Label: Code-Review=-1