Use tripleo-ci-base-standalone parent for podman job
Currently tripleo-ci-rhel-8-standalone-rdo parent is used
and it has yaml anchors defined. But scenario 000 podman
job depends on podman-package-rhel-8 job and got triggered
during changes to podman related files. Due to the yaml
anchors, Zuul is not able to freeze the job graph.

Moving to a seperate parent fixes the issue.

Change-Id: I05b3143fe02883e2682fa85d3d77b91fc9857d2d
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>