Update patch set 11
Patch Set 11: Code-Review-2

I had not realized this change was touching the base job. Doing this is definitely a bad idea, it has the potential to break deployment projects.

If a job needs a package from EPEL, it should take care of enabling the repo as part of the job itself, never modifying the base job.

Patch-set: 11
Label: Code-Review=-2