Update patch set 2
Patch Set 2: Code-Review-1

centos-7 in rdo should be exactly the same as the upstream centos-7 node.  This confuses me as to why it would be deprecated in rdo?

Reuse the DIB from upstream to build the centos-7 node in rdo and it should remain active / supported until upstream deprecates centos-7.

Patch-set: 2
Label: Code-Review=-1