Add tempest image_path override into centos-8 standalone-4
As in related-bug this fails to download the cirros image. Looks
like the override defined in the featureset is ignored because
we have the same override via featureset_override in the job
definition. Adding it there instead and we can remove from fs52
in a followup that depends on this.

Change-Id: If086a9f713c6bdf6937c7e55fd2d7686bca181a7
Related-Bug: 1867807