Add fs039 train job in check
fs039 centos7 job won't run in master because of centos7,
that means it will never run on oooq and oooq-extras repos
because they are branchless and always is master.
Add a specific fs039 job that runs train code in it and can be
triggered on oooq(-extras) repos to check files that are important
for 039 jobs, like freeipa role etc.
Change-Id: I931983df57844cde6e772ae0c1bd40c3f9022825