Run mirror info role in TripleO RDO jobs
This role was initially added to base playbook like
it's done in upstream base job with [1]. For some CentOS8
jobs it was skipped with [2] and [3], but still other
jobs can hit an issue like [4] when mirror is not in sync
with trunk.rdo repos, So adding this role here instead of
base to avoid those issues. This role is needed so RDO mirrors
can be used in jobs that are directly inherited from upstream
jobs.

[1] https://review.rdoproject.org/r/#/c/25870
[2] https://review.rdoproject.org/r/#/c/25900/
[3] https://review.rdoproject.org/r/#/c/25926/
[4] https://bugs.launchpad.net/tripleo/+bug/1869871

Change-Id: I435f3ecb27ba7cfd0aeef0cea44a6be53a536000