Split c7 standalone-full-tempest into two jobs
Standalone-full-tempest job sometimes goes out in timeout state due to
race conditions and concurrency issue.

With this change we are splitting the functionally of
standalone-full-tempest into two new jobs:-

* standalone-full-tempest-api and standalone-full-tempest-scenario

We are not removing the
current(periodic-tripleo-ci-centos-7-standalone-full-tempest-master)
job yet. We will plan to remove it once these new jobs are running
successfully. We have also added these new job definition for
stable/train but currently we have not wired them up into any zuul
layout.

We are testing run of above jobs for stable/train here
https://review.rdoproject.org/r/#/c/26217

Change-Id: I60cd3432d413f7f399d73e540997d1cd5c4bf992