Mark tripleo-ci-centos-8-ovb-3ctlr_1comp_1supp-featureset039 voting
Removes the voting false from centos8 fs39. In the best case it
doesn't matter anyway since this runs in 3rd party, and in the
worst case it is just ignored.

It can still be ignored but perhaps not as easily as if it is
'non voting'. The job was added at [1]

[1] https://tree.taiga.io/project/tripleo-ci-board/task/1604

Change-Id: I8fd01e54357716dc1d23949fd6a02af943b27e04