OVN centos8: add ovn to the enabled providers
OVN jobs are failing tempest test
octavia_tempest_plugin.tests.scenario.v2.test_traffic_ops.TrafficOperationsScenarioTest

With error

"Provider 'ovn' is not enabled."

as tempest configuration is not setting an important value for the test.
This patch adds ovn provider to the list of enabled providers, so the test can run.

Change-Id: Ic78c59883cd43afe1c99ba9041a91e97919bdac4