Update patch set 39
Patch Set 39: -Code-Review

> So... The idea with this patch is to also test latest ansible,
 > right? I'm fine with this, but lacking some context.

Thank you Jpena, John and gfidente for the comments.

During CentOS-8 porting, ceph-ansible was not compatible with ansible-2.9. It got fixed.
Later on we realized that, since we already have jobs for third party testing(where we can test changes using check-rdo or check-github comment on pull requests), we can re-use the same and build the ansible rpm from stable-2.9 branches and test it with ceph-ansible and tripleo so that we can fix the issues early and we have more confidence in our deployment.

In this current job, ansible-2.9 is getting tested with ceph-ansible stable-4.0 with master tripleo. All things are working fine and these jobs are different.

The check-rdo or check-github will trigger these jobs based on master with ceph-ansible stable-4.0
* tripleo-ceph-integration-centos-8-scenario001-standalone
* tripleo-ceph-integration-centos-8-scenario004-standalone

May be we fix this one for train only should be based on centos-7 (as centos-8 tripleo-ci jobs are not yet ready)
Master jobs should use only master branch of ceph-ansible? May be plugging ansible-2.9 will help to stay ahead and fix issues early.

Patch-set: 39