Temp ussuri component *testing* pipeline
Create a placeholder for ussuri component pipeline while don't
have the required artefacts/containers/images to run component
validation jobs like the standalone/scenario ones.

The plan is to enable component-pipeline-ussuri-testing  in a
first moment. Then when we promote to promoted-components and
get all dependencies resolved, we switch to the full template
component-pipeline-ussuri and the -testing one can be deleted.

https://tree.taiga.io/project/tripleo-ci-board/us/1650

Change-Id: I21fe61dc64ea53f30f4bf2ec2307f32ebdda4f64