Mini integration pipeline for ussuri
Create a mini integration pipeline for ussuri jobs
that will build the necessary artefacts for enabling
component pipeline (containres/images).

The mini pipeline contains:
 - contrainer build
 - image build
 - standalone
 - scen001

This pipeine is defined in a template and will temporarily run
in 'openstack-periodic-master' pipeline. The full pipeline will
be created in a second stage w/ all jobs, and renamed to
openstack-periodic-latest-released.

When this happens, we shift all pipelines, moving jobs from:
 - openstack-periodic-master to openstack-periodic-latest-released
 - openstack-periodic-latest-released to openstack-periodic-24hr
 - openstack-periodic-24hr openstack-periodic-wednesday-weekend
 - delete * from openstack-periodic-wednesday-weekend

Change-Id: I7dbb25cc37effa47051bc4bed5888b27bfe87af5