Mini integration pipeline for ussuri
Create a mini integration pipeline for ussuri jobs
that will build the necessary artefacts for enabling
component/integration pipelines (containers/images).

The mini pipeline contains:
 - promote to tripleo-ci-testing
 - standalone
   - container build
 - fs001
   - image build full
   - image build ipa

NOTE: This pipeline is defined in a template and will temporarily
run in 'openstack-periodic-master' pipeline. The full pipeline
will be created in a second stage w/ all jobs, and renamed to
openstack-periodic-latest-released.

Depends-On: https://review.opendev.org/#/c/721842

https://tree.taiga.io/project/tripleo-ci-board/us/1650

Change-Id: I7dbb25cc37effa47051bc4bed5888b27bfe87af5