Update patch set 39
Patch Set 39:

> let me add some "mappings" into the conversation to see if it helps
 > 
 > ceph-ansible stable-3.2 goes with tripleo/queens and ansible 2.7
 > ceph-ansible stable-4.0 goes with tripleo/train and ansible 2.9
 > ceph-ansible stable-4.0 will be used also in tripleo/ussuri
 > 
 > we'll probably never use ceph-ansible stable-5.0 and rather move to
 > a different tool for tripleo/victoria; until then we'll continue to
 > deploy from stable-4.0

Hey guys I'm not sure if I'm saying something stupid
or out of scope here because I missed some context on
the goal of this work.
Basically the value I see running "check-rdo" from a
github PR is that you can do the following actions:

1. clone the current released ceph-ansible upstream (*)
2. checkout the current PR and apply it to the released
content
3. trigger scenario001 and scenario004 to catch issues
or regressions introduced  in this context by the checked-out
PR from the ceph-ansible branch

So I see a real improvement triggering the rdo CI from
github when you're checking out the current PR and testing
the potential code that can be merged and released.
This can help us to prevent code which can be good and can
be merged for ceph-ansible CI point of view, but that can
break one of our use case/scenario(s).


anyway if the purpose here is just to integrate ansible-2.9
in the process, I'm +1 for that.


(*) the released ceph-ansible upstream is the one you can
find here [1] and that is tagged with -release by centos
storage sig [2]; this is the same used by upstream TripleO
CI when scenario00{1,4} are triggered, so basically you
should just clone the ceph-ansible repo passing the "latest"
tag
(e.g., git clone -b v4.0.22 https://github.com/ceph/ceph-ansible.git)



[1] https://github.com/ceph/ceph-ansible/releases
[2] https://cbs.centos.org/koji/buildinfo?buildID=29607

Patch-set: 39
Reviewer: Gerrit User 602 <602@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0