Move stein job to openstack-periodic-wednesday-weekend pipeline
As per the new design [1] moved the stein jobs from
openstack-periodic-24hr to openstack-periodic-wednesday-weekend
pipeline. As we move the stein jobs so only train jobs are
there in integration-pipeline-train-stein.yaml template so rename
that to integration-pipeline-train.yaml and accordingly change the
template name to triggerd the jobs.

So while doing this change, two problems were encountered which are given below:

1. Template name - Currently openstack-periodic-wednesday-weekend pipeline
   contains only rocky and queens jobs, and now this patch is going to add
   stein jobs, accordingly will need to change the template name as the name
   is given on the basis of relases name (integration-pipeline-rocky-queens.yaml).
   To triggered the pipeline will also need to change same in config repo [2].
2. Chiken-egg problem - Doing above mentioned changes in both the repos
   (rdo-jobs and config) simultaneously encountered zuul -1.

So to avoid the template name changes everytime in future,
here will keep standard name integration-pipeline-master-3-4-5 releases
(i.e master-3 -> stein, master-4 -> rocky and master-5 -> queens relases jobs)

And to avoid the chicken-egg problem will complete [1] in 3 steps as given below:

1. In this patch will introduce new template with "integration-pipeline-master-3-4-5"
   and will keep old as is.
2. In the follow patch [2] will remove the old template
3. Add new template name in config repo to triggered the pipeline and
   remove the old one [3]

[1]: https://tree.taiga.io/project/tripleo-ci-board/task/1832?kanban-status=1447274
[2]: https://review.rdoproject.org/r/#/c/28261/
[3]: https://review.rdoproject.org/r/#/c/28212/

Signed-off-by: Bhagyashri Shewale <bshewale@redhat.com>
Change-Id: I6d9f0015a775d6d83d2c788424411755a927f2d5