Set branch for ceph-ansible-integration jobs
When these jobs were created, they did not include any branch limitation.
Now, they are included in the openstack-regular pipeline, which means
that every 4 hours we are executing all those jobs for the 25+ branches
in the github.com/ceph/ceph-ansible repository.

We need to limit the branches it runs on to avoid extra Zuul churn.

Change-Id: Ia1019de5849b91d1d0c63f83b00fff40217f4b72