Tempest skip list standalone scenario jobs
Since Fs021 jobs no longer exists on CentOS-8 and we have
new skip list tool in the town. In order to inhance our
coverage we are running all skip lists tests on different
scenario as each scenarios have different tempest tests.

It will help to keep the coverage in order to findout the
passing skip list tests.

It will be triggered in weekend pipeline.

Change-Id: I35376f165ed4bc3c331c95e95b39f03939e9e08a
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>