Fix train to ussuri upgrade job
We switched to pacemaker enabled in Ussuri, so by default the upgrade
from Train -> Ussuri is non-HA to HA. To test HA always in the upgrade
we passes docker-ha.yaml environment file via featureset file[1].

We need to include the podman.yaml for centos8 jobs when defining
docker-ha.yaml because in train docker-ha.yaml was docker based.

Periodic job definition for upgrade is missing standalone_environment_files
in comparision to upstream job definition. With this patch we are syncing
periodic job definition to upstream upgrade job definition[2]

[1] https://opendev.org/openstack/tripleo-quickstart/src/branch/master/config/general_config//featureset056.yml#L27
[2] https://opendev.org/openstack/tripleo-ci/src/branch/master/zuul.d/standalone-jobs.yaml#L162-L165

Change-Id: Ie4738d1ec6e703eb2218c319ed7a0a4ac67acc17
Closes-Bug: lp#1900957