Switch rdopkg-reqcheck job to a Fedora 32 node
The rdopkg-reqcheck job was using a now obsolete version of Fedora (31),
and it was supposed to migrate to the container-fedora nodeset some time ago.

This commit updates the container-fedora nodeset to use Fedora 33, although
it still requires some work before container images can be used. For now,
it updates the job to use a Fedora 32 node.

Change-Id: I0a1e13a43c53caf7985e92d5834ab1340c137427