[wip] Add sc10 ovn in network component
Third party Load Balancer driver can be integrated with Octavia.
Functionality related to this has been developed in OVN and OVN
can now be supported as a provider driver for Octavia.

ovn-octavia-provider comes from network component so it make sense
to add this job in network component.

We will check further on existing plain sc10 coverage scope in network
component and if that can be removed after adding this job.

Change-Id: I33d6eff817fb27cd3218c7506d7454af7507ec16