Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> The periodic job has been fairly stable in the past days, passing all tempest tests. I think testing more than smoke tests in RDO jobs provide a great value and avoids downstream CIX escalations. I am voting CR-1 just to get a better understanding of what motivated this change that significantly reduces test coverage.
> 
> https://review.rdoproject.org/zuul/builds?job_name=periodic-tripleo-ci-centos-8-scenario010-standalone-master

I'm definitely not an Octavia expert but we have for pretty long time opened CIX card https://trello.com/c/UFBhp5oz which is related to timeouts in that specific job. As per feedback from Brian and Gregory which I had my understanding was that we should run just smoke tests there as running all Octavia tests is too much for that job.
If job is stable now, maybe we can simply close that CIX card and abandon that patch. I'm totally fine with that. Please just provide some feedback in the CIX card if we can close it or what we should do still to close it finally.

Patch-set: 3
Reviewer: Gerrit User 534 <534@270e2033-b340-4cff-9539-693957ebf0e7>