Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I am sorry but I don't quite follow yet the reason for this change since the job has been mostly stable looking at https://review.rdoproject.org/zuul/builds?job_name=periodic-tripleo-ci-centos-8-scenario010-standalone-master&project=openstack%2Ftripleo-ci
> 
> Of course I am only one and non-core, so my vote weighs in little. I won't take offense if this patch is approved, but I cannot upvote something I don't agree with :)

The job was stable because the failing smoke test was disabled, https://review.opendev.org/c/openstack/openstack-tempest-skiplist/+/778235 is the patch to enable that. If we want to run whole test, we can bump the job timing. Then we will be good?

Patch-set: 3
Reviewer: Gerrit User 111 <111@270e2033-b340-4cff-9539-693957ebf0e7>