Create new jobs template deps-distgit-check-jobs for dependencies
This review is part of the implementation of a new process to
manage dependencies in RDO.

With new process, a distgit will be maintained in review.r.o for each
dependency. This patch is creating a new template that will be used to
gate these new repos. As first job, i'm adding deps-validate-cbs which
do a scratch build in CBS from the distgit content.

Depends-On: https://review.rdoproject.org/r/c/config/+/33501
Change-Id: I5bb816d0aa87f42041ee84fcc52c214bab94899b