Fix undercloud upgrade job definition and parent
In upstream undercloud-upgrade job has parent
pointing to tripleo-ci-base-singlenode-centos-8[1]
and in rdo side, It is using standalone upgrade parent
job leading to failing upgrade job.

The parent needs to be the same as upstream which
is getting modified to tripleo-ci-base-singlenode-rdo-centos-8[2]
in rdo.

This patch fixes the same.

It also fixes the job definition keeping in sync with
upstream [3]

[1]. https://opendev.org/openstack/tripleo-ci/src/branch/master/zuul.d/deprecated-jobs.yaml#L228
[2]. https://review.rdoproject.org/r/gitweb?p=config.git;a=blob;f=zuul.d/tripleo-rdo-base.yaml#l180
[3]. https://opendev.org/openstack/tripleo-ci/src/branch/master/zuul.d/deprecated-jobs.yaml#L227

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id82c84241c63a291a30bfc09f41e60e4b77f1d96