Switch to UEFI boot on featureset001 master
On master featureset001 jobs, the OVB baremetal_image_name is set to
ipxe-uefi-boot, which is identical to ipxe-boot except the glance
property hw_firmware_type has value 'uefi'.

Any cloud running these jobs will need the ipxe-uefi-boot image as
well as the ipxe-boot image. This has been done on vexxhost.

This results in the generated instackenv.json to have boot_mode:uefi.
This allows the job to test UEFI boot with ironic provisioning, which
will soon be the default.

Change-Id: I386f869c154eb8ad5b27b3c0538cc8465f7c842b