Update patch set 1
Patch Set 1: Code-Review-1

+1 to sandeep points

as we discussed on the last scrum this will have to be done on a per branch basis i.e. for master its the ones here including the comments from ysandeep - as commented for some of those we should include them on the older branches.

I guess it will be easier to have the conversation this way anyway so thank you for splitting (i.e. having one review per branch so we can get the list of services right should be easier)

Patch-set: 1
Reviewer: Gerrit User 192 <192@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1