OVB 3ctlr 1comp component job for VF testing
Executes validations role, testing the full VF cli and features implemented.
Furthemore the executes list of validations, representing minimal viable sample
of the VF validations catalogue.

The validations are executed in a single batch, with any number failing causing
the job itself to fail.

Job is defined for the master and Wallaby branches only.
Expansion to other branches should be undertaken only after substantial
consideration of the costs and benefits.

Depends-On: https://review.opendev.org/c/openstack/tripleo-quickstart-extras/+/815292
Depends-On: https://review.opendev.org/c/openstack/validations-common/+/817041
Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/817392

Co-Authored-by: matbu <mat.bultel@gmail.com>

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I3d59c50671028cb5b5aec5f6b9424456ad685fd9