Edit TARGET envvar
In [1], we check if TARGET starts with "centos9", so we need to change
the $TARGET here. Plus, this is better for code consistency even if
centos-stream-9 would be better with regard to name convention.
We hit issue in centos9-xena since it's the first non-master release
tested (i.e master follow another path execution).

[1] https://github.com/softwarefactory-project/DLRN/blob/master/scripts/run_tests.sh#L81

Change-Id: I634b57069953590d72e6cbf9b9d640963887c40b