[centos-opstools] Move/replace opstools jobs from rdo/config
Putting them in the same place makes it easier to debug (less places to
look for job definitions) and means that changes to jobs can be tested
(via speculative merging) if a CI failure is caused by the job rather
than a change in the project being tested.

Add in job to indicate unsupported packages

This job will be attached to deprecated packages in opstools, so that
change requests will be marked as Verified-1, and the author will be
warned that the package is not supported or built anymore.

This allows for a host of jobs to be cleaned up that run against these
projects.

Change-Id: I961c7613f9cb0fe4bb7808cc39ffa6d931613761