[opstools] Create opstools-deprecated-warning job
This job is intended to warn developer that the package is no longer
maintained. It serves to help the opstools developer grok which projects
are active or not, and helps to document the state of package support.

This job will be attached to deprecated packages in opstools, so that
change requests will be marked as Verified-1, and the author will be
warned that the package is not supported or built anymore.

This allows for a host of jobs to be cleaned up that run against these
projects.

Change-Id: I5a13604a54d3ad4cc9c7c39fe07793ad0239f4ac