[opstools] Create new ``centos-opstools-*-distgit-cbs`` jobs
Having the opstools jobs in the same place makes it easier to debug (less places to
look for job definitions) and means that changes to jobs can be tested
(via speculative merging) if a CI failure is caused by the job rather
than a change in the project being tested.

* pre.yaml is included in the ``centos-opstools-distgit-cbs-base`` job in rdo-config
* the jobs are run on centos-8-stream, because the reqs aren't on
  centos9stream
* The parameters for pkg-export and pkg-check are constructed from the distro
  var, which means that child jobs can be added for different distros by
  modifying this var in the job definition::
    vars:
      distro: '8s'
* the job includes a ``buildroot`` var, which defaults to ``collectd-5``,
  and can be modified in the future, if we require jobs which do the same
  thing except with different buildroots (e.g. ``collectd-6`` (future),
  or ``perfmon`` (use in the past))

This makes future updates easier to grok and manage.

Depends-On: https://review.rdoproject.org/r/c/config/+/38317
Change-Id: I961c7613f9cb0fe4bb7808cc39ffa6d931613761