[opstools] Create new ``opstools-*-distgit-cbs`` jobs
Having the opstools jobs in the same place makes it easier to debug (less places to
look for job definitions) and means that changes to jobs can be tested
(via speculative merging) if a CI failure is caused by the job rather
than a change in the project being tested.

* pre.yaml is included in the ``opstools-distgit-cbs-base`` job in rdo-config
* The parameters for pkg-export and pkg-check can be constructed from
  the distro information, allowing the job's ``nodeset`` parameter to
  determine the param, which means that child jobs can be added for
  different distros by modifying the ``nodeset``.

This makes future updates easier to grok and manage.

Depends-On: https://review.rdoproject.org/r/c/config/+/38317
Change-Id: I961c7613f9cb0fe4bb7808cc39ffa6d931613761