Make non-voting yoga weirdo jobs on cs8 for rdoinfo
We are blocked at this point as those jobs deploy puppet modules from
rpms while puppet-promotion uses them from code, so our puppet-promoted
packages don't pass these jobs. Also, the promotion pipeline is merged
because of the version of testtools we have in deps.

In order to unblock the promotion of new testtools i'm setting weirdo
jobs as non-voting. Note that, in parallel, i'm testing the testtools
update with the consistent repo.

Change-Id: I2e5e04b99a1c1060d91cf0298442625ba9f3de41