Add rpmlint exception for octavia
Upstream review [1] has introduced cmd/prometheus_proxy.py. Although the
executable is created via console_scripts, a shebang has been added to
that file (unnecessarily) so rpmlint is complaining about a file with
shebang which doesn't have execution perms. As this should be fixed
upstream, i'm adding an exception to the rpmlint job.

[1] https://review.opendev.org/c/openstack/octavia/+/812258/19/octavia/cmd/prometheus_proxy.py

Change-Id: If3bc09b48a388e6479aefb4aed123f09360ade43