Switch to Zed tags for weirdo jobs
master release switched to Zed in rdoinfo with [1],
this patch switches weirdo jobs setting Zed as master. Note that zed
will not support CS8 so we are only creating jobs to gate changes in
zed-testing for CS9 with weirdo.

Also, it creates the weirdo jobs required to gate updates to
yoga-release.

[1] https://review.rdoproject.org/r/c/rdoinfo/+/41314

Change-Id: I75707d8531f0b5c0f6564ca82b9c93690b943b76