Remove tripleo-ci-base-standalone-periodic-centos-8 base job
tripleo-ci-base-standalone-periodic-centos-8 is the old parent
and will be removed via this job
https://review.rdoproject.org/r/c/config/+/44455
and it needs to be replaced with
periodic-tripleo-ci-standalone-rdo-centos-8.

It is used everywhere as a part of new zuul job parenting.
It also removes unwanted ceph and podman jobs from projects.yaml.
These jobs are no longer maintained. ansible-collection, ceph-ansible,
podman jobs will be removed in a separate review.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I7b60c51585f9392b06968cd4a608dd4af159a41e