Move wallaby c8 jobs to run on vexx
The quay-push-wallaby and push-wallaby centos 8 jobs are timeing out a
lot running on ibm cloud, and with the lack of logs, it is very hard to
identify the root cause right now. In order to not block the entire
pipeline, we are reverting it to run on vexx, while we debug the root
cause. Running these jobs manually on vexx are not failing at the
moment.

Change-Id: Ia9487bb330d20adc83972a75e2e936389f44bf52
Related-Bug: #1988810