[ovb] Add openstack/requirements under required-projects
Currently in ovb-manage role, we are using constraint_file
from opendev.org/openstack/requirements.

In CI, while pulling the upper-constraints.txt file from
opendev.org also leads to Network is unreachable in CI run.

And the job wents to Retry Limit.

For fixing it, we are adding openstack/requirements under
ovb required-projects job so that it gets clonned in the
CI job workspace. We can use the upper-constraints.txt file
from there.

https://review.rdoproject.org/r/c/testproject/+/45294/2#message-a230c36a8479bf81956c6d9d862c4de147be6174
already tests the same.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Ibaba49eeae5ea965a19c1b3805aa5cad05f51338