Ignore TripleO jobs for other centos versions in CBS updates
When running tripleo to gate changes in CBS tags we need to select the
jobs not only based on phase (testing or release), version of openstack
but also based on version of CentOs Stream. i.e. if we are updating
nfv9s-openvswitch-2-testing, jobs for CS8 should be skipped.

This review is improving the initial check where we find if a job is
required or can be ignored to include this criteria.

Change-Id: I85fc7e14f6d7062f66b521bec50df01ec53de92d