Stop setting mixed_os_stable_version for centos 9 mixed OS jobs
The mixed_os_stable_version was used as part of the fix for [1]
to ensure correct delorean reporting. However after [2] and the
addition of [3] this parameter is now only needed/used by the
centos 8 *component* job. This patch is needed by [4] where we
simplify the fix for [1] by keeping only the c8 case.

[1] https://bugs.launchpad.net/tripleo/+bug/1990012
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/46119
[3] https://review.rdoproject.org/r/c/config/+/46130
[4] https://review.rdoproject.org/r/c/config/+/46161

Change-Id: I05a623314d65b47dbce12a6da40f12516a79d691