Teardown wallaby c8 integration and component jobs and pipelines
After [1][2] we are running the required centos8/wallaby jobs
(containers, ffu and mixed-rhel) as part of the centos9
wallaby line. This patch removes the centos8 integration
and component pipeline as well as related job definitions eg
tempest-skiplist-jobs, except a subset of jobs we are keeping
(container builds and ffu jobs). Needed by [3] as depends-on.

[1] https://review.rdoproject.org/r/c/rdo-jobs/+/46119
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/46147
[3] https://review.rdoproject.org/r/c/config/+/46399

Depends-On: https://review.rdoproject.org/r/c/config/+/46400

Change-Id: I8fc0b6c725cb8fa716a1b1019d8aaab26529f69c